Optimize code

+++ crates/batata-core/src/service/distro.rs
@@ -660,7 +660,7 @@ impl DistroProtocol {
                     use std::hash::{Hash, Hasher};
                     let mut hasher = std::collections::hash_map::DefaultHasher::new();
                     std::time::SystemTime::now().hash(&mut hasher);
-                    (hasher.finish() % 2000) as u64
+                    hasher.finish() % 2000
                 };
                 tokio::time::sleep(config.verify_interval + Duration::from_millis(jitter_ms)).await;
